Restructuring and moving sources around: moved all convert-related sources to convert/ dir
