+++ kguitar/convert/convertascii.cpp
