+++ kguitar/convert/convertascii.h
