+++ kguitar/convert/convertbase.h
