+++ kguitar/convert/convertgtp.cpp
