+++ kguitar/convert/convertmidi.cpp
