+++ kguitar/convert/converttse3.cpp
