+++ kguitar/kguitar_part.cpp
@@ -11,13 +11,13 @@
 #include "trackdrag.h"
 #include "settings.h"
 
-#include "convertkg.h"
-#include "convertascii.h"
-#include "convertxml.h"
-#include "convertmidi.h"
-#include "converttse3.h"
-#include "converttex.h"
-#include "convertgtp.h"
+#include "convert/convertkg.h"
+#include "convert/convertascii.h"
+#include "convert/convertxml.h"
+#include "convert/convertmidi.h"
+#include "convert/converttse3.h"
+#include "convert/converttex.h"
+#include "convert/convertgtp.h"
 
 #include "optionsexportascii.h"
 #include "optionsexportmusixtex.h"
